requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sqlalchemy==2.0.30
python-dotenv==1.0.1
google-generativeai==0.5.4
passlib==1.7.4
bcrypt==4.1.3
itsdangerous==2.2.0
jinja2==3.1.4
python-multipart==0.0.9
stripe==9.8.0
requests==2.32.3